move authors to blog-authors component
