+++ app/controllers/posts.js
@@ -3,6 +3,6 @@ import Ember from "ember";
 export default Ember.Controller.extend({
   page: 1,
   perPage: 5,
-  
-  queryParams: ["page", "perPage", "query"],
+
+  queryParams: ["page", "perPage", "query"]
 });
