+++ tests/integration/components/blog-authors-test.js
@@ -1,0 +1,27 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+
+moduleForComponent('blog-authors', 'Integration | Component | blog authors', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  assert.expect(2);
+
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{blog-authors}}`);
+
+  assert.equal(this.$().text(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#blog-authors}}
+      template block text
+    {{/blog-authors}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
